Adapters: Removed the shared memory hang-monitoring
This is because of a problem caused by the system's select() or poll() which removes the timeout for all other processes listening on the same socket when one of the processes' select()/poll() call returns. The removal of the timeout will make it seem like all the other processes has hanged (unless they get a new request before the original timeout should have timed out) and will force the managing process to restart all but one of the children.
This will result in a continual restart loop as long as there are few requests per time unit, if there are none or many this will not occur though because the timeouts will work normally or because for each request the processes notify that they have not yet hanged.